Use pipe delimited messages
Previously we used space delimited messages, and attempted to quote
some of the fields when they might contain a space as well. Instead
we'll use a pipe character (e.g. |) because that should not appear
in any of our fields except *maybe* user agent, but that is the last
field and we can safely handle it.