Stop logging unknown UAs to Sentry
It turns out that sentry is not a particularly great data store for
our unknown User Agents. Since I haven't been looking at this list
recently, and the inclusion of them is preventing other projects from
getting meaningful data out of Sentry, we're going to just stop doing
that.

Eventually we should re-add some mechanism for tracking what unknown
UAs we're getting so that we can correctly handle them.